+++ src/main/resources/static/sql/table.sql
@@ -1,5 +1,5 @@
 create table member(
-    id bigint auto_increment primary key,
+    id int auto_increment primary key,
     name varchar(255) not null unique,
     password varchar(255) not null
 );
